Update data extraction and UI code, add database functions
The changes made include updating the data extraction methods in vote_parser.py and modifying the user interface code in ok_vote_ui.py. The methods for extracting data from doc files was commented out and will be replaced with the process of pulling data from tables to store as CSV. Additionally, key database functions were incorporated in db_config.py to facilitate the export of vote details and roll calls to CSV along with an insert function for roll calls.